Button: update to barrel
