+++ .eslintrc
@@ -133,7 +133,8 @@
         "no-use-before-define": "off",
         "import/no-extraneous-dependencies": "off",
         "no-unused-vars": "off",
-        "react/require-default-props": "off"
+        "react/require-default-props": "off",
+        "no-restricted-exports": "off"
       }
     },
     {
