+++ client/common/Button/Button.stories.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 
-import Button from './Button';
-import { GithubIcon, DropdownArrowIcon, PlusIcon } from './icons';
+import Button from './index';
+import { GithubIcon, DropdownArrowIcon, PlusIcon } from '../icons';
 
 export default {
   title: 'Common/Button',
